Did some clean up to the P9FSPrim implementation: 1) removed all but cmd and msg file 2) made both cmd and message treat open, read, write, close    as a session where the data produced as a side effect of a    write is cached in the fid associated with the open the    data is freed on a close.
This is not throughly tested but does seem to have the same
behaviour as before.  Need to go over this with Eric and test
that malloc/frees are are paired and working.

Calls via ixp don't seem to work (eg.  I am not seeing
the klunks and the associated frees.