Large set of changes sorry!  New Event interface, Event Driven Console test app, bug fixes and more.
New Event Interface to the EventManager.  You can now specify an arbitrary EBBID and method
to be dispatched when and event occurs.

Changes to the Event Manager and pic to allow an event to be enabled and disabled
via vecon vecoff pic interfaces.  Which globally enable and disable the vector.
The enable and disable code needs to be cleaned up.

Fixed bugs that Dan pointed out in pic dispatching... pselects interrupted
by signals can not rely on fd sets being valid must exit and restart select.

Enriched unix lrt_pic_src to be a data structure which can specify a fd and flags
that indicate what kind of events should be monitored on the fd.  This permits
a device to specify that an FD should be monitored for read, write or error independently.

Added a simple kludgy event driven console app using the above infrastructure.

cleanup some inconsistencies in EBB creation and malloc.