fixed race conditions on msgmgr startup and retry message send if target msgmgr not up
Made a whole series of changes to carefully make sure that the rep for msgmgr that the
root knows about  on a event location is not valid until the event handler and rep are
all intialized.  Then, the sender gets back an error on find target, and can loop until
the message mgr is up.