fixed comment to be consistent with Dan's change in where bit set
The issue is, if we have an RCU mechanism, where we want to drain references
on all nodes where a object has been accessed, the current bitvector will
not let you know all the accesses.  Thats fine for now, but, wanted to document
it, and we may eventually want to add an accessed bit vector as well.