Fix numerous baremetal bugs
All tests seem to be working now. There were several subtle bugs that
required some modifications.

- The event location is assumed to be linear for a lot of our code but
  the apic id is not linear, so some mapping had to be done.

- There was an issue where we forgot to zero out the IDT (It was
  switched from bss to allocated recently). On hardware this caused a
  general protection fault that was quite hard to track down

The outstanding concern is with the ioapic which isn't being
initialized. It should initialize to reasonable defaults (no IRQ
routing). The issue is that this IO APIC has a different version than
the one I setup with, so I am apprehensive about just initializing it
the same way as the other version.